Using an NPM github module.  no more branches.
